File Restructure in Components/Views
