+++ src/frontend/src/views/MyTeam/MyTeamPage.jsx
@@ -14,23 +14,23 @@ import {
   TableCell,
   Paper,
 } from "@mui/material";
-import NavBar from "../components/NavBar";
-import RosterTable from "../components/Tables/RosterTable";
-import ScheduleTable from "../components/Tables/ScheduleTable";
-import NotificationsRow from "../components/Rows/NotificationsRow";
-import { useAuth } from "../hooks/AuthProvider";
-import { getPlayerById } from "../api/player";
-import { getNotificationsByTeamId } from "../api/notification";
+import NavBar from "../../components/NavBar";
+import RosterTable from "../../components/Tables/RosterTable";
+import ScheduleTable from "../../components/Tables/ScheduleTable";
+import NotificationsRow from "../../components/Rows/NotificationsRow";
+import { useAuth } from "../../hooks/AuthProvider";
+import { getPlayerById } from "../../api/player";
+import { getNotificationsByTeamId } from "../../api/notification";
 import { useEffect, useState } from "react";
 import TabContext from "@mui/lab/TabContext";
 import TabList from "@mui/lab/TabList";
 import TabPanel from "@mui/lab/TabPanel";
-import NoDataCard from "../components/Cards/NoDataCard";
+import NoDataCard from "../../components/Cards/NoDataCard";
 import { useNavigate, useParams } from "react-router-dom";
-import { getScheduleGamesByTeamId } from "../api/team";
-import LoadingOverlay from "../components/LoadingOverlay";
-import ContactInfoTable from "../components/Tables/ContactInfoTable";
-import { removePlayerFromRoster } from "../api/team";
+import { getScheduleGamesByTeamId } from "../../api/team";
+import LoadingOverlay from "../../components/LoadingOverlay";
+import ContactInfoTable from "../../components/Tables/ContactInfoTable";
+import { removePlayerFromRoster } from "../../api/team";
 
 // McMaster colours - AI Generated
 const MCMASTER_COLOURS = {
