+++ src/frontend/src/views/Schedule/TeamScheduleView.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
-import { useAuth } from "../hooks/AuthProvider";
-import { getPlayerById } from "../api/player";
-import { getScheduleBySeasonId } from "../api/schedule";
-import { formatDate } from "../utils/Formatting";
-import { Container, Typography, Box } from "@mui/material";
+import { useAuth } from "../../hooks/AuthProvider";
+import { getPlayerById } from "../../api/player";
+import { getScheduleGamesByTeamId } from "../../api/team";
+import { formatDate } from "../../utils/Formatting";
+import ReschedulePopup from "../Reschedule/ReschedulePopup";
+import { getAvailableGameslots } from "../../api/reschedule-requests";
 
-// Helper to normalize a date to local ISO (YYYY-MM-DD)
 const getLocalISODate = (date) => {
   const d = new Date(date);
   d.setHours(12, 0, 0, 0);
@@ -14,7 +14,6 @@ const getLocalISODate = (date) => {
   return localDate.toISOString().split("T")[0];
 };
 
-// Add McMaster colours constant
 const MCMASTER_COLOURS = {
   maroon: '#7A003C',
   grey: '#5E6A71',
@@ -22,7 +21,6 @@ const MCMASTER_COLOURS = {
   lightGrey: '#F5F5F5',
 };
 
-// Common loading overlay style for all schedule components
 const loadingOverlayStyle = {
   position: 'absolute',
   top: '50%',
@@ -39,43 +37,54 @@ const loadingOverlayStyle = {
   zIndex: 1000,
 };
 
-export const LeagueSchedule = () => {
+export const TeamSchedule = () => {
   const [currentMonth, setCurrentMonth] = useState(new Date());
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const auth = useAuth();
   const [player, setPlayer] = useState(null);
-  const [seasonGames, setSeasonGames] = useState([]);
+  const [teamGames, setTeamGames] = useState([]);
+  const [showModal, setShowModal] = useState(false);
+  const [selectedDate, setSelectedDate] = useState("");
+  const [selectedMatch, setSelectedMatch] = useState("");
+  const [availableGameslots, setAvailableGameslots] = useState({});
 
-  // Helper functions defined before use
+  // Move getMonthDates function before it's used
   const getMonthDates = (date) => {
     let firstDayOfMonth = new Date(date.getFullYear(), date.getMonth(), 1);
     let lastDayOfMonth = new Date(date.getFullYear(), date.getMonth() + 1, 0);
-    let firstWeekday = firstDayOfMonth.getDay();
+    let firstWeekday = firstDayOfMonth.getDay(); // 0 (Sun) to 6 (Sat)
     let totalDays = lastDayOfMonth.getDate();
     let daysArray = [];
+    // Add empty cells for days before the first day
     for (let i = 0; i < firstWeekday; i++) {
       daysArray.push(null);
     }
+    // Add each day as an ISO date string
     for (let d = 1; d <= totalDays; d++) {
       daysArray.push(formatDate(new Date(date.getFullYear(), date.getMonth(), d)));
     }
     return daysArray;
   };
 
-  // Combine data fetching into a single useEffect
+  // Now we can safely use getMonthDates
+  const monthYear = currentMonth.toLocaleDateString("en-US", { month: "long", year: "numeric" });
+  const monthDates = getMonthDates(currentMonth);
+
+  // Combine all data fetching into a single useEffect
   useEffect(() => {
     const fetchAllData = async () => {
       try {
         setLoading(true);
-        
+        setError(null);
+
         // Fetch player data
         const playerData = await getPlayerById(auth.playerId);
         setPlayer(playerData.player);
 
-        // Fetch season games if player has a team with seasonId
-        if (playerData.player?.team?.seasonId) {
-          const scheduleData = await getScheduleBySeasonId(playerData.player.team.seasonId.id);
+        // Fetch team games if player has a team
+        if (playerData.player?.team?.id) {
+          const gamesData = await getScheduleGamesByTeamId(playerData.player.team.id);
           
           // Filter games for current month
           const startOfMonth = new Date(currentMonth.getFullYear(), currentMonth.getMonth(), 1);
@@ -83,40 +92,54 @@ export const LeagueSchedule = () => {
           const startISO = getLocalISODate(startOfMonth);
           const endISO = getLocalISODate(endOfMonth);
 
-          const filteredGames = scheduleData.schedule.games.filter((game) => {
-            const gameDate = getLocalISODate(new Date(game.date));
+          const filteredGames = gamesData.games.filter((game) => {
+            const gameDate = getLocalISODate(game.date);
             return gameDate >= startISO && gameDate <= endISO;
           });
           
-          setSeasonGames(filteredGames);
+          setTeamGames(filteredGames);
         }
+
+        // Fetch available gameslots
+        const slotsResponse = await getAvailableGameslots();
+        const formattedData = slotsResponse.reduce((acc, slot) => {
+          const dateKey = formatDate(new Date(slot.date));
+          const slotString = `${slot.time} | ${slot.field}`;
+          if (!acc[dateKey]) acc[dateKey] = [];
+          acc[dateKey].push({ id: slot._id, slotString: slotString });
+          return acc;
+        }, {});
+        
+        setAvailableGameslots(formattedData);
+
       } catch (err) {
-        console.error(err.message || "Failed to fetch data");
+        setError(err.message || "Failed to fetch data");
       } finally {
         setLoading(false);
       }
     };
 
     fetchAllData();
-  }, [auth.playerId, currentMonth]);
-
-  const monthYear = currentMonth.toLocaleDateString("en-US", { month: "long", year: "numeric" });
-  const monthDates = getMonthDates(currentMonth);
+  }, [auth.playerId, currentMonth]); // Only depend on playerId and currentMonth
 
   // Navigation: change month
   const handleNavigation = (direction) => {
-    setLoading(true);
-    let newDate = new Date(currentMonth);
+    const newDate = new Date(currentMonth);
     newDate.setMonth(newDate.getMonth() + (direction === "next" ? 1 : -1));
     setCurrentMonth(newDate);
   };
 
-  // For each day, filter all season games (normalized) that match this day
+  // Add handler for reschedule clicks
+  const handleRescheduleClick = (dateKey, match) => {
+    setSelectedDate(dateKey);
+    setSelectedMatch(match);
+    setShowModal(true);
+  };
+
+  // Modify the getMatchesForDay function to include the reschedule button
   const getMatchesForDay = (dayISO) => {
     if (!dayISO) return null;
-    const matches = seasonGames.filter(
-      (game) => formatDate(new Date(game.date)) === dayISO
-    );
+    const matches = teamGames.filter((game) => formatDate(game.date) === dayISO);
     if (matches.length === 0) return null;
     return matches.map((match, idx) => (
       <div key={idx} style={styles.matchText}>
@@ -124,6 +147,14 @@ export const LeagueSchedule = () => {
         <div style={{ fontSize: '11px', marginTop: '2px', color: MCMASTER_COLOURS.maroon }}>
           {match.time} | {match.field}
         </div>
+        {player?.team?.captainId?.id === player._id && (
+          <button
+            style={styles.rescheduleButton}
+            onClick={() => handleRescheduleClick(dayISO, match)}
+          >
+            Reschedule
+          </button>
+        )}
       </div>
     ));
   };
@@ -160,13 +191,9 @@ export const LeagueSchedule = () => {
         )}
 
         {!loading && (
-          <div style={{
-            ...styles.calendar,
-          }}>
+          <div style={styles.calendar}>
             {["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"].map((day) => (
-              <div key={day} style={{
-                ...styles.calendarHeader,
-              }}>
+              <div key={day} style={styles.calendarHeader}>
                 {day}
               </div>
             ))}
@@ -187,6 +214,24 @@ export const LeagueSchedule = () => {
           </div>
         )}
       </div>
+
+      {/* ReschedulePopup modal */}
+      {showModal && (
+        Object.keys(availableGameslots).length > 0 ? (
+          <ReschedulePopup
+            selectedDate={selectedDate}
+            selectedMatch={selectedMatch}
+            availableTimeslots={availableGameslots}
+            player={player}
+            onClose={() => setShowModal(false)}
+          />
+        ) : (
+          <div style={loadingOverlayStyle}>
+            <div style={styles.spinner}></div>
+            <p>Loading available timeslots...</p>
+          </div>
+        )
+      )}
     </>
   );
 };
@@ -198,12 +243,14 @@ const styles = {
     alignItems: "center",
     width: '100%',
   },
+  
   title: {
     fontSize: "24px",
     fontWeight: "bold",
     textAlign: "center",
     color: MCMASTER_COLOURS.maroon,
   },
+  
   navButton: {
     backgroundColor: MCMASTER_COLOURS.maroon,
     color: "white",
@@ -211,10 +258,12 @@ const styles = {
     padding: "10px 15px",
     borderRadius: "5px",
     cursor: "pointer",
+    transition: 'background-color 0.2s ease',
     '&:hover': {
       backgroundColor: '#5A002C',
     }
   },
+  
   calendar: {
     display: "grid",
     gridTemplateColumns: "repeat(7, 1fr)",
@@ -225,6 +274,7 @@ const styles = {
     overflow: 'hidden',
     boxShadow: '0 2px 12px rgba(0,0,0,0.05)',
   },
+  
   calendarHeader: {
     backgroundColor: MCMASTER_COLOURS.maroon,
     color: "white",
@@ -234,6 +284,7 @@ const styles = {
     fontSize: '0.9rem',
     letterSpacing: '0.5px',
   },
+  
   calendarCell: {
     backgroundColor: 'white',
     padding: '12px 8px',
@@ -245,22 +296,26 @@ const styles = {
       backgroundColor: MCMASTER_COLOURS.lightGrey,
     },
   },
+  
   emptyCell: {
     backgroundColor: '#f8f8f8',
     minHeight: '120px',
   },
+  
   dateText: {
     fontWeight: "500",
     fontSize: "14px",
     color: MCMASTER_COLOURS.maroon,
     marginBottom: '8px',
   },
+  
   matchesContainer: {
     display: 'flex',
     flexDirection: 'column',
     gap: '6px',
     marginTop: '4px',
   },
+  
   matchText: {
     fontSize: '12px',
     color: MCMASTER_COLOURS.grey,
@@ -277,6 +332,25 @@ const styles = {
       boxShadow: '0 3px 6px rgba(0,0,0,0.08)',
     },
   },
+  
+  rescheduleButton: {
+    backgroundColor: MCMASTER_COLOURS.gold + '90',
+    color: MCMASTER_COLOURS.maroon,
+    border: 'none',
+    padding: '6px 10px',
+    borderRadius: '4px',
+    fontWeight: '500',
+    cursor: 'pointer',
+    fontSize: '11px',
+    marginTop: '6px',
+    transition: 'all 0.2s ease',
+    '&:hover': {
+      backgroundColor: MCMASTER_COLOURS.gold,
+      transform: 'translateY(-1px)',
+      boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
+    },
+  },
+  
   spinner: {
     width: '30px',
     height: '30px',
@@ -287,4 +361,4 @@ const styles = {
   },
 };
 
-export default LeagueSchedule;
+export default TeamSchedule;
